Move lib/api-keys and platform/api-keys into infra/auth
Change-type: patch
